+++ src/main/java/net/silentchaos512/mechanisms/block/alloysmelter/AlloySmelterTileEntity.java
@@ -77,10 +77,6 @@ public class AlloySmelterTileEntity extends AbstractMachineTileEntity<AlloySmelt
     @SuppressWarnings("AssignmentOrReturnOfFieldWithMutableType")
     @Override
     public int[] getSlotsForFace(Direction side) {
-        if (side == Direction.UP)
-            return SLOTS_INPUT;
-        if (side == Direction.DOWN)
-            return SLOTS_OUTPUT;
         return SLOTS_ALL;
     }