+++ src/main/java/net/silentchaos512/mechanisms/block/compressor/CompressorTileEntity.java
@@ -68,10 +68,6 @@ public class CompressorTileEntity extends AbstractMachineTileEntity<CompressingR
     @SuppressWarnings("AssignmentOrReturnOfFieldWithMutableType")
     @Override
     public int[] getSlotsForFace(Direction side) {
-        if (side == Direction.UP)
-            return SLOTS_INPUT;
-        if (side == Direction.DOWN)
-            return SLOTS_OUTPUT;
         return SLOTS_ALL;
     }