+++ src/main/java/net/silentchaos512/mechanisms/block/electricfurnace/ElectricFurnaceTileEntity.java
@@ -71,10 +71,6 @@ public class ElectricFurnaceTileEntity extends AbstractMachineTileEntity<Abstrac
     @SuppressWarnings("AssignmentOrReturnOfFieldWithMutableType")
     @Override
     public int[] getSlotsForFace(Direction side) {
-        if (side == Direction.UP)
-            return SLOTS_INPUT;
-        if (side == Direction.DOWN)
-            return SLOTS_OUTPUT;
         return SLOTS_ALL;
     }