+++ CHANGELOG.md
@@ -4,6 +4,8 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [0.7.5] - 2020-07-02
+### Changed
+- Machines can now have items inserted and extracted from any side
 ### Fixed
 - Iron/gold smelting/blasting recipes missing [#90]