+++ CHANGELOG.md
@@ -3,6 +3,10 @@
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [0.7.6]
+### Changed
+- Server config is now common config [#94]
+
 ## [0.7.5] - 2020-07-02
 ### Changed
 - Machines can now have items inserted and extracted from any side